[#39] Move schedule_callable() to llm_wrapper, leading to further simplification
